+++ src/main/kotlin/com/github/copyinaction/config/jwt/JwtAuthenticationFilter.kt
@@ -24,9 +24,13 @@ class JwtAuthenticationFilter(
     }
 
     private fun resolveToken(request: HttpServletRequest): String? {
-        val bearerToken = request.getHeader("Authorization")
-        if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7)
+        val cookies = request.cookies
+        if (cookies != null) {
+            for (cookie in cookies) {
+                if (cookie.name == "accessToken") {
+                    return cookie.value
+                }
+            }
         }
         return null
     }